# Documentacion/openapi/full_api.yaml
openapi: 3.0.3
info:
  title: Plataforma - API (Auth, Invoices, OCR, SUNAT)
  description: >-
    API completa y profesional para la plataforma de facturación, OCR y
    verificación SUNAT. Incluye contratos para Auth, Invoices (CRUD,
    upload-pdf, apply-parsed, verify-ruc), OCR ingestion/status/suggestions y
    servicio SUNAT.
  version: 1.0.0
servers:
  - url: https://api.example.com/api
    description: URL base (reemplazar en cada entorno)
  - url: http://localhost:5000/api
    description: Entorno de desarrollo (usar el servidor local apropiado)
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Iniciar sesión (devuelve accessToken y opcional refresh cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refrescar access token (acepta cookie httpOnly o body)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refrescado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Cerrar sesión (revoca refresh token)
      responses:
        '204':
          description: Logout realizado
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/me:
    get:
      tags: [auth]
      summary: Obtener información del usuario actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
...
security:
  - bearerAuth: []